+++ CHANGELOG.md
@@ -327,3 +327,4 @@
 - [2026-04-11 23:03] fix(security): prevent double tap duplicate action in security
 - [2026-04-12 09:05] fix(forms): prevent double tap duplicate action in forms
 - [2026-04-12 09:38] fix(storage): fix memory leak by cancelling storage subscription
+- [2026-04-12 09:44] test(network): add unit tests for network business logic